LFSR encryption and decryption based projects
These were part of a set of questions for an interview. The first one named "TextFileEncryptor" is a simple text file encryptor/decryptor I created that utilizes an LFSR-based algorithm. The second one named "CustomKBDFileDecryptor" is a decryptor/encryptor for a custom file format that was given to me to decrypt. That is the program that decrypts and encrypts it.